requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.89.0
uvicorn[standard]==0.20.0
uvloop == 0.17.0
httptools == 0.5.0
gunicorn==20.1.0
mysql-connector-python == 8.0.33
python-decouple==3.3
passlib == 1.7.4
pydantic == 1.10.7
uuid == 1.30
PyJWT==2.7.0

